% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonotoneHR.R
\docType{data}
\name{chernoff_realizations}
\alias{chernoff_realizations}
\title{A dataframe containing computed chernoff distribution}
\format{
A data frame with 201 rows and 3 variables:
\describe{
  \item{xcoor}{domain on which chernoff distribution if computed}
  \item{DF}{Distribution function of the Chernoff distribution}
  \item{density}{Density of the Chernoff distribution}
}
}
\usage{
chernoff_realizations
}
\description{
A dataframe containing computed chernoff distribution
}
\keyword{datasets}
