% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonotoneHR.R
\name{gcm.unique}
\alias{gcm.unique}
\title{A function used to remove the repeated numbers for the computation of gcm/lcm.}
\usage{
gcm.unique(x, y)
}
\arguments{
\item{x}{A vector corresponds to the independent variable.}

\item{y}{A vector corresponds to the dependent variable.}
}
\value{
\code{logcm} GCM of the curve y~x.
}
\description{
A function used to remove the repeated numbers for the computation of gcm/lcm.
}
