% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonotoneHR.R
\name{NA.est}
\alias{NA.est}
\title{A function used to construct the Nelson-Aalen estimator.}
\usage{
NA.est(surv.data)
}
\arguments{
\item{surv.data}{A dataframe containing right-censored survival time and censor status.}
}
\value{
\code{est} The Nelson-Aalen object.
}
\description{
A function used to construct the Nelson-Aalen estimator.
}
