% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonotoneHR.R
\docType{data}
\name{survData}
\alias{survData}
\title{A right-censored example data}
\format{
A data frame with 1217 rows and 3 variables:
\describe{
  \item{time}{right-censored time}
  \item{status}{object censored or not, with status=0 implies being censored.}
  \item{group}{Treatment group of the objects. By default, for a non-decreasing hazard ratio, S is the group on the numerator and T is the group on the denominator.}
}
}
\usage{
survData
}
\description{
A dataset containing the right-censored survival time and censor status for two groups of objects
}
\keyword{datasets}
